Adds explicit cache control headers to bucket objects
In order to workaround the problems with Cloudfront
caching the 3xx responses necessary for this whole dance
to work, the default TTL on the Cloudfront drstribution was
lowered to 0. Each individual item in the bucket now gets an
explicit Cache Control header to ensure we still get edge
caching.

1 month was chosen as this is the current default.

Good explanation of the caching/redirect loop problem outlined
in this thread:
https://stackoverflow.com/questions/42308291/resize-images-on-the-fly-in-cloudfront-and-get-them-in-the-same-url-instantly-a