Add support for GIFs
Turns out a bunch of folks have uploaded GIFs for profile pictures.
Sharp image library doesn't support outputtng GIFs, so I've added a hack
that outputs it as a PNG in the destination bucket, with an image/png
mimeType, but preserving the original filename.

We probably don't want GIFs anyway on the site, worth preventing users
uploading them in the future. Meanwhile, this workaround will allow the
CDN to serve the existing ones.